Move deposit to block_operations
